+++ classes/block.py
@@ -7,11 +7,11 @@ _COMMENT_START = '#'
 
 _MULTILINE_STRING_ERROR = "Multiline string"
 
-class Section:
-    """A Section is a collection of lines (strings) in a filter, which may include rules in them.
+class Block:
+    """A block is a collection of lines (strings) in a filter, which may include rules in them.
     
-    Lines in a section may be appended, removed, commeted out or swapped for other lines.
-    Sections may have rules which can be obtained from this class.
+    Lines in a block may be appended, removed, commeted out or swapped for other lines.
+    Blocks may have rules which can be obtained from this class.
     Rules are automatically extracted from lines when new lines are added.
     """
 
@@ -22,29 +22,29 @@ class Section:
 
     @classmethod
     def extract(cls, lines: list[str], line_count: int = 1):
-        "Returns all sections in a from a list of text lines."
-        sections: list[Section] = []
-        current = Section(line_count)
+        "Returns all blocks in a from a list of text lines."
+        blocks: list[Block] = []
+        current = Block(line_count)
         for line in lines:
-            #check if the line starts with show or hide, if so start a new section
+            #check if the line starts with show or hide, if so start a new block
             line_stripped = line.lstrip()
             if line_stripped.startswith(_SHOW) or line_stripped.startswith(_HIDE):
                 if not current.is_empty():
-                    sections.append(current)
-                current = Section(line_count)
+                    blocks.append(current)
+                current = Block(line_count)
             current.append(line.rstrip())
             line_count +=1
-        #this is the last section in the file and is omitted in the loop
+        #this is the last block in the file and is omitted in the loop
         if not current.is_empty():
-            sections.append(current)
-        return sections
+            blocks.append(current)
+        return blocks
 
     def is_empty(self):
-        """Returns true if the section has no lines. False if it has any."""
+        """Returns true if the block has no lines. False if it has any."""
         return len(self.lines) == 0
 
     def append(self, line: str):
-        """Appends the line at the end of the section. If any rules are present, they are extracted. Multiline strings are not allowed."""
+        """Appends the line at the end of the block. If any rules are present, they are extracted. Multiline strings are not allowed."""
         line_number = self.line_number + len(self.lines)
         if '\n' in line:
             raise GeneratorError(_MULTILINE_STRING_ERROR, line_number)
@@ -103,7 +103,7 @@ class Section:
                 self.lines[i] = line
     
     def hide(self):
-        """Attempts to completely hide a section by setting it to 'Hide' and
+        """Attempts to completely hide a block by setting it to 'Hide' and
         commenting out the lines that show effects on the screen if they are present."""
         self.swap("Show", "Hide")
         self.comment("PlayAlertSoundPositional")
@@ -111,7 +111,7 @@ class Section:
         self.comment("PlayEffect")
     
     def show(self):
-        """Attempts to show a section fully, by setting it to 'Show' and un-commenting out
+        """Attempts to show a block fully, by setting it to 'Show' and un-commenting out
         the lines that show effects on the screen if they are present.
         This is the reverse function to hide."""
         self.swap("Hide", "Show")
@@ -120,7 +120,7 @@ class Section:
         self.uncomment("PlayEffect")
 
     def get_rules(self, rule_name: str):
-        """Gets all the rules in the section with name equals to rule_name."""
+        """Gets all the rules in the block with name equals to rule_name."""
         return [ rule for rule in self.rules if rule.name == rule_name ]
                     
     def __str__(self):