+++ classes/filter.py
@@ -1,7 +1,7 @@
 import os
 
 from classes.generator_error import GeneratorError
-from .section import Section
+from .block import Block
 
 _FILE_EXISTS_ERROR = "The file path corresponds to an already existing file"
 _FILE_NOT_FOUND_ERROR = "The input file was not found"
@@ -10,15 +10,15 @@ _PERMISSION_ERROR = "You don't have permission to read or write on this director
 class Filter:
     """The filter class is a representation of a .filter file."""
 
-    def __init__(self, filepath: str, sections: list[Section] = None):
+    def __init__(self, filepath: str, blocks: list[Block] = None):
         self.filepath: str = filepath
-        self.sections: list[Section] = sections or self._get_sections(filepath)
+        self.blocks: list[Block] = blocks or self._get_blocks(filepath)
 
-    def _get_sections(self, filepath: str):
+    def _get_blocks(self, filepath: str):
         try:
             with open(self.filepath, "r") as file:
                 lines = file.readlines()
-                return Section.extract(lines)
+                return Block.extract(lines)
         except FileNotFoundError:
             raise GeneratorError(_FILE_NOT_FOUND_ERROR, filepath=filepath)
         except PermissionError:
@@ -29,8 +29,8 @@ class Filter:
         self._create_directory(output_filepath)
         try:
             with open(output_filepath, "w") as file:
-                for section in self.sections:
-                    for line in section.lines:
+                for block in self.blocks:
+                    for line in block.lines:
                         file.write(line + '\n')
         except FileExistsError:
             raise GeneratorError(_FILE_EXISTS_ERROR, filepath=self.filepath)