+++ classes/generator.py
@@ -11,15 +11,15 @@ class Generator:
     The generate function loads a module with the 'handler' name and calls a function 'handle' on it. It is up to the handler any processing required."""
 
     def generate(self, filter: Filter, handler_name: str, options: list[str]):
-        """Imports the module specified by handler and gives it a list of sections, a filepath to output its result and a list of option strings. The rest is up to the handler.
+        """Imports the module specified by handler and gives it a reference to the filter, a Block and a list of option strings. The rest is up to the handler.
         The filepath must be writable.
         The handler must be a python file under the directory "handlers", and the '.py' extension must be omitted.
-        The handler must include the function handle(filter: Filter, section: Section, options: list[str])."""
+        The handler must include the function handle(filter: Filter, block: Block, options: list[str])."""
         handler = self._get_handler(handler_name)
-        new_sections = []
-        for section in filter.sections:
-            new_sections += handler.handle(filter, section, options)
-        return Filter(filter.filepath, new_sections)
+        new_blocks = []
+        for block in filter.blocks:
+            new_blocks += handler.handle(filter, block, options)
+        return Filter(filter.filepath, new_blocks)
 
     def _get_handler(self, handler_name: str):
         try: