+++ handlers/import.py
@@ -2,22 +2,22 @@ import os.path
 
 from classes.filter import Filter
 from classes.rule import Rule
-from classes.section import Section
+from classes.block import Block
 from common import util
 
 _NAME = "import"
 
-def handle(filter: Filter, section: Section, _):
-    """Handles imports of other filter files' sections. Options are ignored."""
-    return _import_sections(section, [ filter.filepath ])
+def handle(filter: Filter, block: Block, _):
+    """Handles imports of other filter files' blocks. Options are ignored."""
+    return _import_blocks(block, [ filter.filepath ])
 
-def _import_sections(section: Section, filepaths: list[str]):
-    new_sections = [ section ]
-    for rule in section.get_rules(_NAME):
+def _import_blocks(block: Block, filepaths: list[str]):
+    new_blocks = [ block ]
+    for rule in block.get_rules(_NAME):
         filter = _load_filter(rule, filepaths)
-        for section in filter.sections:
-            new_sections += _import_sections(section, filepaths + [ filter.filepath ])
-    return new_sections
+        for block in filter.blocks:
+            new_blocks += _import_blocks(block, filepaths + [ filter.filepath ])
+    return new_blocks
 
 def _load_filter(rule: Rule, filepaths: list[str]):
     filter_filepath = util.parse_rule_filepath(filepaths[-1], rule.description)